+++ src/components/App/components/public/Connectors/Connector/Fields/Content/index.js
@@ -33,7 +33,7 @@ class Content extends EditorComponent {
 
   getRootElement() {
 
-    return "span";
+    return "div";
   }